+++ test/deploy.js
@@ -81,6 +81,7 @@ test('deploy required args', async t => {
       maxFeePerGas: undefined,
       maxPriorityFeePerGas: undefined
     },
+    metadata: undefined,
   });
 });
 
@@ -100,6 +101,7 @@ test('deploy all args', async t => {
       '--gasPrice', '1000000000', // 1 gwei
       '--maxFeePerGas', '2000000000', // 2 gwei
       '--maxPriorityFeePerGas', '500000000', // 0.5 gwei
+      '--metadata', '{ "commitHash": "4ae3e0d", "tag": "v1.0.0", "anyOtherField": "anyValue" }',
     ];
 
   await deploy(args, t.context.fakeDeployClient, t.context.fakeNetworkClient);
@@ -122,6 +124,24 @@ test('deploy all args', async t => {
       gasPrice: '0x3b9aca00',
       maxFeePerGas: '0x77359400',
       maxPriorityFeePerGas: '0x1dcd6500',
+    },
+    metadata: {
+      commitHash: '4ae3e0d',
+      tag: 'v1.0.0',
+      anyOtherField: 'anyValue',
     }
   });
+});
+
+test('deploy with invalid metadata JSON', async t => {
+  const args = [
+      '--contractName', 'MyContract',
+      '--contractPath', 'contracts/MyContract.sol',
+      '--chainId', FAKE_CHAIN_ID,
+      '--buildInfoFile', 'test/input/build-info.json',
+      '--metadata', 'v1.0.0', // not valid JSON
+    ];
+
+  const error = await t.throwsAsync(deploy(args, t.context.fakeDeployClient, t.context.fakeNetworkClient));
+  t.true(error.message.includes('Failed to parse metadata option as JSON'), error.message);
 });